+++ CHANGELOG.md
@@ -6,6 +6,7 @@
    ### Changed
       - more container quality of life config auto detections
       - progress meter also logs throughput to file log on tty runs
+      - emit version on dump / restore one shots
       - (couch) additional tracing output
       - (couch) chunk large operations
       - (influx2) additional debugging and tracing output